+++ BeRich/BeRichApp.swift
@@ -1,10 +1,3 @@
-//
-//  BeRichApp.swift
-//  BeRich
-//
-//  Created by Danila on 11.04.2023.
-//
-
 import SwiftUI
 
 @main
