+++ BeRich/Extensions/Color+Extension.swift
@@ -1,10 +1,3 @@
-//
-//  Color+Extension.swift
-//  BeRich
-//
-//  Created by Ilya Solovyov on 13.04.2023.
-//
-
 import SwiftUI
 
 extension Color {
