+++ BeRich/Extensions/View+Extension.swift
@@ -1,10 +1,3 @@
-//
-//  View+Extension.swift
-//  BeRich
-//
-//  Created by Ilya Solovyov on 13.04.2023.
-//
-
 import SwiftUI
 
 public extension View {
