+++ BeRich/ListScreen/ListScreen.swift
@@ -1,25 +1,19 @@
-//
-//  ListScreen.swift
-//  BeRich
-//
-//  Created by Danila on 11.04.2023.
-//
-
 import SwiftUI
 
 struct ListScreen: View {
-    @State private var tickers: [Ticker] = Fakes.tickersFakes
+    private let navigationTitle = "BeRich"
+    @State private var tickers: [Ticker] = Fakes.tickers
 
     var body: some View {
         NavigationStack {
             List(tickers) { ticker in
                 TickerCellView(ticker: ticker)
-                    .listRowInsets(EdgeInsets())
+                    .listRowInsets(EdgeInsets(.zero))
                     .listRowSeparator(.hidden)
             }
             .scrollContentBackground(.hidden)
             .background(Color.background)
-            .navigationTitle("BeRich")
+            .navigationTitle(navigationTitle)
             .navigationBarColor(
                 backgroundColor: Color.blueMain,
                 titleColor: Color.white
