+++ BeRich/ListScreen/Ticker.swift
@@ -1,28 +1,4 @@
-//
-//  Ticker.swift
-//  BeRich
-//
-//  Created by Ilya Solovyov on 13.04.2023.
-//
-
 import SwiftUI
-
-enum PriceChange {
-    case increase(Double)
-    case stable
-    case decrease(Double)
-
-    var color: Color {
-        switch self {
-        case .increase:
-            return .greenMain
-        case .stable:
-            return .gray400
-        case .decrease:
-            return .redMain
-        }
-    }
-}
 
 struct Ticker: Identifiable {
     let id: String = UUID().uuidString
